fullScreenMod
I took the pickaday.js calendar and tweeked it open a more elegant,
full screen calendar. I also added a few cool styling changes, .SVG
custom arrows and changes some of the inline .js to reflect date
specific requirements. It’s not fully responsive yet, but that will be
in the next commit.